requirements.txt
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
PyYAML==5.3.1
coverage==5.3.1
django-cors-headers==3.7.0
requests==2.25.1
